fix: keep Effect tripwire out of the frontend typecheck
src/ tests cannot import Node builtins under the Vite/Svelte tsconfig.
Move the Effect import walk next to the TypeScript slot check, pin
bun.lock @typescript/old to real TypeScript 6, and name the boundary
scanner as a TS7.0-window stand-in rather than an AST parser to finish.
